requirements.txt
beautifulsoup4>=4.12.2
numpy>=1.19.5
pandas>=1.1.5
psycopg2-binary>=2.9
python-dotenv>=0.19.0
requests>=2.27.0
SQLAlchemy>=1.4.45

